refactor: rename test helper macro test_base_store to test_recall_by_id
All tests macros actually test storage as a side effect, the main
 difference with this test is that it recalls by id, so name changed to
 make it match the pseudo-naming-scheme
